# storage/scribe/openapi.yaml
openapi: 3.0.3
info:
  title: 'Evolução local de snipe-it API Documentation'
  description: 'This documentation aims to provide all the information you need to work with the Snipe-IT API.'
  version: 1.0.0
servers:
  -
    url: 'https://snipe-it.test'
...
paths:
...
  '/api/v1/categories/{item_type}/selectlist':
    get:
      summary: Selectlist
      operationId: selectlist
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Unauthorized or unauthenticated.'
                properties:
                  error:
                    type: string
                    example: 'Unauthorized or unauthenticated.'
      tags:
        - Categories
    parameters:
      -
        in: path
        name: item_type
        description: ''
        example: architecto
        required: true
        schema:
          type: string
...
  /api/v1/departments:
    get:
      summary: 'List Departments'
      operationId: listDepartments
      description: ''
      parameters:
        -
          in: query
          name: search
          description: 'Search term to filter results.'
          example: IT
          required: false
          schema:
            type: string
            description: 'Search term to filter results.'
            example: IT
            nullable: false
        -
          in: query
          name: name
          description: 'Filter by exact department name.'
          example: IT
          required: false
          schema:
            type: string
            description: 'Filter by exact department name.'
            example: IT
            nullable: false
        -
          in: query
          name: company_id
          description: 'Filter by exact company ID.'
          example: 1
          required: false
          schema:
            type: integer
            description: 'Filter by exact company ID.'
            example: 1
            nullable: false
        -
          in: query
          name: manager_id
          description: 'Filter by exact manager (user) ID. Example:'
          example: 16
          required: false
          schema:
            type: integer
            description: 'Filter by exact manager (user) ID. Example:'
            example: 16
            nullable: false
        -
          in: query
          name: location_id
          description: 'Filter by exact location ID.'
          example: 1
          required: false
          schema:
            type: integer
            description: 'Filter by exact location ID.'
            example: 1
            nullable: false
        -
          in: query
          name: sort
          description: 'Column to sort results by. Allowed values: id, name, image, users_count, notes, created_at, updated_at, location, manager, company. Default: created_at.'
          example: name
          required: false
          schema:
            type: string
            description: 'Column to sort results by. Allowed values: id, name, image, users_count, notes, created_at, updated_at, location, manager, company. Default: created_at.'
            example: name
            nullable: false
        -
          in: query
          name: order
          description: 'Order of sorted results. Allowed values: asc, desc. Default: desc.'
          example: asc
          required: false
          schema:
            type: string
            description: 'Order of sorted results. Allowed values: asc, desc. Default: desc.'
            example: asc
            nullable: false
        -
          in: query
          name: offset
          description: 'Offset/starting position of the results. Default: 0.'
          example: 0
          required: false
          schema:
            type: integer
            description: 'Offset/starting position of the results. Default: 0.'
            example: 0
            nullable: false
        -
          in: query
          name: limit
          description: 'Limit the number of results returned. Default: 25. Maximum: 100.'
          example: 50
          required: false
          schema:
            type: integer
            description: 'Limit the number of results returned. Default: 25. Maximum: 100.'
            example: 50
            nullable: false
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Unauthorized or unauthenticated.'
                properties:
                  error:
                    type: string
                    example: 'Unauthorized or unauthenticated.'
      tags:
        - Departments
    post:
      summary: 'Create Department'
      operationId: createDepartment
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Departments
...